+++ packages/google_mobile_ads/android/src/main/java/io/flutter/plugins/googlemobileads/FlutterAdManagerInterstitialAd.java
@@ -21,6 +21,7 @@ import com.google.android.gms.ads.LoadAdError;
 import com.google.android.gms.ads.admanager.AdManagerInterstitialAd;
 import com.google.android.gms.ads.admanager.AdManagerInterstitialAdLoadCallback;
 import com.google.android.gms.ads.admanager.AppEventListener;
+import java.lang.ref.WeakReference;
 
 /**
  * Wrapper around {@link com.google.android.gms.ads.admanager.AdManagerInterstitialAd} for the
@@ -42,10 +43,12 @@ class FlutterAdManagerInterstitialAd extends FlutterAd.FlutterOverlayAd {
    * null only until `load` is called.
    */
   public FlutterAdManagerInterstitialAd(
+      int adId,
       @NonNull AdInstanceManager manager,
       @NonNull String adUnitId,
       @NonNull FlutterAdManagerAdRequest request,
       @NonNull FlutterAdLoader flutterAdLoader) {
+    super(adId);
     this.manager = manager;
     this.adUnitId = adUnitId;
     this.request = request;
@@ -58,30 +61,22 @@ class FlutterAdManagerInterstitialAd extends FlutterAd.FlutterOverlayAd {
         manager.activity,
         adUnitId,
         request.asAdManagerAdRequest(),
-        new AdManagerInterstitialAdLoadCallback() {
-          @Override
-          public void onAdLoaded(@NonNull AdManagerInterstitialAd adManagerInterstitialAd) {
-            FlutterAdManagerInterstitialAd.this.ad = adManagerInterstitialAd;
-            ad.setOnPaidEventListener(
-                new FlutterPaidEventListener(manager, FlutterAdManagerInterstitialAd.this));
-            ad.setAppEventListener(
-                new AppEventListener() {
-                  @Override
-                  public void onAppEvent(@NonNull String name, @NonNull String data) {
-                    manager.onAppEvent(FlutterAdManagerInterstitialAd.this, name, data);
-                  }
-                });
-            manager.onAdLoaded(
-                FlutterAdManagerInterstitialAd.this, adManagerInterstitialAd.getResponseInfo());
-          }
+        new DelegatingAdManagerInterstitialAdCallbacks(this));
+  }
 
-          @Override
-          public void onAdFailedToLoad(@NonNull LoadAdError loadAdError) {
-            manager.onAdFailedToLoad(
-                FlutterAdManagerInterstitialAd.this, new FlutterLoadAdError(loadAdError));
-            super.onAdFailedToLoad(loadAdError);
-          }
-        });
+  void onAdLoaded(AdManagerInterstitialAd ad) {
+    this.ad = ad;
+    ad.setAppEventListener(new DelegatingAdManagerInterstitialAdCallbacks(this));
+    ad.setOnPaidEventListener(new FlutterPaidEventListener(manager, this));
+    manager.onAdLoaded(adId, ad.getResponseInfo());
+  }
+
+  void onAdFailedToLoad(LoadAdError loadAdError) {
+    manager.onAdFailedToLoad(adId, new FlutterLoadAdError(loadAdError));
+  }
+
+  void onAppEvent(@NonNull String name, @NonNull String data) {
+    manager.onAppEvent(adId, name, data);
   }
 
   @Override
@@ -90,7 +85,47 @@ class FlutterAdManagerInterstitialAd extends FlutterAd.FlutterOverlayAd {
       Log.e(TAG, "The interstitial wasn't loaded yet.");
       return;
     }
-    ad.setFullScreenContentCallback(new FlutterFullScreenContentCallback(manager, this));
+    ad.setFullScreenContentCallback(new FlutterFullScreenContentCallback(manager, adId));
     ad.show(manager.activity);
+  }
+
+  @Override
+  void dispose() {
+    ad = null;
+  }
+
+  /**
+   * This class delegates various rewarded ad callbacks to FlutterAdManagerInterstitialAd. Maintains
+   * a weak reference to avoid memory leaks.
+   */
+  private static final class DelegatingAdManagerInterstitialAdCallbacks
+      extends AdManagerInterstitialAdLoadCallback implements AppEventListener {
+
+    private final WeakReference<FlutterAdManagerInterstitialAd> delegate;
+
+    DelegatingAdManagerInterstitialAdCallbacks(FlutterAdManagerInterstitialAd delegate) {
+      this.delegate = new WeakReference<>(delegate);
+    }
+
+    @Override
+    public void onAdLoaded(@NonNull AdManagerInterstitialAd interstitialAd) {
+      if (delegate.get() != null) {
+        delegate.get().onAdLoaded(interstitialAd);
+      }
+    }
+
+    @Override
+    public void onAdFailedToLoad(@NonNull LoadAdError loadAdError) {
+      if (delegate.get() != null) {
+        delegate.get().onAdFailedToLoad(loadAdError);
+      }
+    }
+
+    @Override
+    public void onAppEvent(@NonNull String name, @NonNull String data) {
+      if (delegate.get() != null) {
+        delegate.get().onAppEvent(name, data);
+      }
+    }
   }
 }