+++ packages/google_mobile_ads/android/src/main/java/io/flutter/plugins/googlemobileads/GoogleMobileAdsViewFactory.java
@@ -57,15 +57,17 @@ final class GoogleMobileAdsViewFactory extends PlatformViewFactory {
   @Override
   public PlatformView create(Context context, int viewId, Object args) {
     final Integer adId = (Integer) args;
-    final PlatformView view = (PlatformView) manager.adForId(adId);
-
-    if (view == null) {
+    FlutterAd ad = manager.adForId(adId);
+    if (ad == null || ad.getPlatformView() == null) {
       return getErrorView(context, adId);
     }
-
-    return view;
+    return ad.getPlatformView();
   }
 
+  /**
+   * Returns an ErrorView with a debug message for debug builds only. Otherwise just returns an
+   * empty PlatformView.
+   */
   private static PlatformView getErrorView(@NonNull final Context context, int adId) {
     final String message =
         String.format(