+++ packages/google_mobile_ads/android/src/test/java/io/flutter/plugins/googlemobileads/FlutterAdManagerInterstitialAdTest.java
@@ -15,6 +15,7 @@
 package io.flutter.plugins.googlemobileads;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
 import static org.mockito.ArgumentCaptor.forClass;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
@@ -63,7 +64,7 @@ public class FlutterAdManagerInterstitialAdTest {
     when(mockFlutterRequest.asAdManagerAdRequest()).thenReturn(mockRequest);
     flutterAdManagerInterstitialAd =
         new FlutterAdManagerInterstitialAd(
-            mockManager, "testId", mockFlutterRequest, mockFlutterAdLoader);
+            1, mockManager, "testId", mockFlutterRequest, mockFlutterAdLoader);
   }
 
   @Test
@@ -100,8 +101,7 @@ public class FlutterAdManagerInterstitialAdTest {
             any(AdManagerInterstitialAdLoadCallback.class));
 
     FlutterLoadAdError flutterLoadAdError = new FlutterLoadAdError(loadAdError);
-    verify(mockManager)
-        .onAdFailedToLoad(eq(flutterAdManagerInterstitialAd), eq(flutterLoadAdError));
+    verify(mockManager).onAdFailedToLoad(eq(1), eq(flutterLoadAdError));
   }
 
   @Test
@@ -152,8 +152,8 @@ public class FlutterAdManagerInterstitialAdTest {
             eq(mockRequest),
             any(AdManagerInterstitialAdLoadCallback.class));
 
+    verify(mockManager).onAdLoaded(1, responseInfo);
     verify(mockAdManagerAd).setOnPaidEventListener(any(FlutterPaidEventListener.class));
-    verify(mockManager).onAdLoaded(flutterAdManagerInterstitialAd, responseInfo);
     final ArgumentCaptor<FlutterAdValue> adValueCaptor = forClass(FlutterAdValue.class);
     verify(mockManager).onPaidEvent(eq(flutterAdManagerInterstitialAd), adValueCaptor.capture());
     assertEquals(adValueCaptor.getValue().currencyCode, "Dollars");
@@ -179,9 +179,10 @@ public class FlutterAdManagerInterstitialAdTest {
     verify(mockAdManagerAd).setFullScreenContentCallback(any(FullScreenContentCallback.class));
     verify(mockAdManagerAd).show(mockManager.activity);
     verify(mockAdManagerAd).setAppEventListener(any(AppEventListener.class));
-    verify(mockManager).onAdShowedFullScreenContent(eq(flutterAdManagerInterstitialAd));
-    verify(mockManager).onAdImpression(eq(flutterAdManagerInterstitialAd));
-    verify(mockManager).onAdDismissedFullScreenContent(eq(flutterAdManagerInterstitialAd));
+    verify(mockManager).onAdShowedFullScreenContent(eq(1));
+    verify(mockManager).onAdImpression(eq(1));
+    verify(mockManager).onAdDismissedFullScreenContent(eq(1));
+    assertNull(flutterAdManagerInterstitialAd.getPlatformView());
   }
 
   @Test
@@ -219,8 +220,7 @@ public class FlutterAdManagerInterstitialAdTest {
         .setFullScreenContentCallback(any(FullScreenContentCallback.class));
 
     flutterAdManagerInterstitialAd.show();
-    verify(mockManager)
-        .onFailedToShowFullScreenContent(eq(flutterAdManagerInterstitialAd), eq(adError));
+    verify(mockManager).onFailedToShowFullScreenContent(eq(1), eq(adError));
   }
 
   @Test
@@ -260,6 +260,6 @@ public class FlutterAdManagerInterstitialAdTest {
     flutterAdManagerInterstitialAd.load();
 
     verify(mockAdManagerAd).setAppEventListener(any(AppEventListener.class));
-    verify(mockManager).onAppEvent(eq(flutterAdManagerInterstitialAd), eq("test"), eq("data"));
+    verify(mockManager).onAppEvent(eq(1), eq("test"), eq("data"));
   }
 }