+++ packages/google_mobile_ads/android/src/test/java/io/flutter/plugins/googlemobileads/FlutterBannerAdTest.java
@@ -15,6 +15,7 @@
 package io.flutter.plugins.googlemobileads;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.mockito.ArgumentCaptor.forClass;
 import static org.mockito.ArgumentMatchers.any;
@@ -35,6 +36,7 @@ import com.google.android.gms.ads.AdView;
 import com.google.android.gms.ads.LoadAdError;
 import com.google.android.gms.ads.ResponseInfo;
 import io.flutter.plugin.common.BinaryMessenger;
+import io.flutter.plugin.platform.PlatformView;
 import io.flutter.plugins.googlemobileads.FlutterAd.FlutterLoadAdError;
 import org.junit.Before;
 import org.junit.Test;
@@ -67,7 +69,7 @@ public class FlutterBannerAdTest {
     doReturn(mockAdView).when(bannerAdCreator).createAdView();
     flutterBannerAd =
         new FlutterBannerAd(
-            mockManager, "testId", mockFlutterRequest, mockFlutterAdSize, bannerAdCreator);
+            1, mockManager, "testId", mockFlutterRequest, mockFlutterAdSize, bannerAdCreator);
   }
 
   @Test
@@ -95,7 +97,7 @@ public class FlutterBannerAdTest {
     verify(mockAdView).setAdUnitId(eq("testId"));
     verify(mockAdView).setAdSize(adSize);
     FlutterLoadAdError expectedError = new FlutterLoadAdError(loadError);
-    verify(mockManager).onAdFailedToLoad(eq(flutterBannerAd), eq(expectedError));
+    verify(mockManager).onAdFailedToLoad(eq(1), eq(expectedError));
   }
 
   @Test
@@ -140,27 +142,34 @@ public class FlutterBannerAdTest {
     verify(mockAdView).setAdListener(any(AdListener.class));
     verify(mockAdView).setAdUnitId(eq("testId"));
     verify(mockAdView).setAdSize(adSize);
+    verify(mockManager).onAdLoaded(eq(1), eq(responseInfo));
+    verify(mockManager).onAdImpression(eq(1));
+    verify(mockManager).onAdClosed(eq(1));
+    verify(mockManager).onAdOpened(eq(1));
+    assertEquals(flutterBannerAd.getPlatformView().getView(), mockAdView);
     verify(mockAdView).setOnPaidEventListener(any(FlutterPaidEventListener.class));
-    verify(mockManager).onAdLoaded(eq(flutterBannerAd), eq(responseInfo));
-    verify(mockManager).onAdImpression(eq(flutterBannerAd));
-    verify(mockManager).onAdClosed(eq(flutterBannerAd));
-    verify(mockManager).onAdOpened(eq(flutterBannerAd));
     final ArgumentCaptor<FlutterAdValue> adValueCaptor = forClass(FlutterAdValue.class);
     verify(mockManager).onPaidEvent(eq(flutterBannerAd), adValueCaptor.capture());
     assertEquals(adValueCaptor.getValue().currencyCode, "Dollars");
     assertEquals(adValueCaptor.getValue().precisionType, 1);
     assertEquals(adValueCaptor.getValue().valueMicros, 1000L);
-    assertEquals(flutterBannerAd.getView(), mockAdView);
   }
 
   @Test
-  public void destroy() {
+  public void dispose() {
     flutterBannerAd.load();
 
-    assertEquals(flutterBannerAd.getView(), mockAdView);
+    assertEquals(flutterBannerAd.getPlatformView().getView(), mockAdView);
+    PlatformView platformView = flutterBannerAd.getPlatformView();
+    assertNotNull(platformView);
 
-    flutterBannerAd.destroy();
+    flutterBannerAd.dispose();
     verify(mockAdView).destroy();
-    assertNull(flutterBannerAd.getView());
+    assertNull(flutterBannerAd.getPlatformView());
+    // Check that the platform view still retains a reference to the view until
+    // dispose is called on it.
+    assertNotNull(platformView.getView());
+    platformView.dispose();
+    assertNull(platformView.getView());
   }
 }