+++ packages/google_mobile_ads/android/src/test/java/io/flutter/plugins/googlemobileads/FlutterInterstitialAdTest.java
@@ -15,6 +15,7 @@
 package io.flutter.plugins.googlemobileads;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
 import static org.mockito.ArgumentCaptor.forClass;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
@@ -63,7 +64,8 @@ public class FlutterInterstitialAdTest {
     when(mockFlutterRequest.asAdRequest()).thenReturn(mockAdRequest);
 
     flutterInterstitialAd =
-        new FlutterInterstitialAd(mockManager, "testId", mockFlutterRequest, mockFlutterAdLoader);
+        new FlutterInterstitialAd(
+            1, mockManager, "testId", mockFlutterRequest, mockFlutterAdLoader);
   }
 
   @Test
@@ -103,7 +105,7 @@ public class FlutterInterstitialAdTest {
             any(InterstitialAdLoadCallback.class));
 
     FlutterLoadAdError expectedError = new FlutterLoadAdError(loadAdError);
-    verify(mockManager).onAdFailedToLoad(eq(flutterInterstitialAd), eq(expectedError));
+    verify(mockManager).onAdFailedToLoad(eq(1), eq(expectedError));
   }
 
   @Test
@@ -152,8 +154,8 @@ public class FlutterInterstitialAdTest {
             eq(mockAdRequest),
             any(InterstitialAdLoadCallback.class));
 
+    verify(mockManager).onAdLoaded(eq(1), eq(responseInfo));
     verify(mockAd).setOnPaidEventListener(any(FlutterPaidEventListener.class));
-    verify(mockManager).onAdLoaded(eq(flutterInterstitialAd), eq(responseInfo));
 
     final ArgumentCaptor<FlutterAdValue> adValueCaptor = forClass(FlutterAdValue.class);
     verify(mockManager).onPaidEvent(eq(flutterInterstitialAd), adValueCaptor.capture());
@@ -179,9 +181,10 @@ public class FlutterInterstitialAdTest {
     flutterInterstitialAd.show();
     verify(mockAd).setFullScreenContentCallback(any(FullScreenContentCallback.class));
     verify(mockAd).show(eq(mockManager.activity));
-    verify(mockManager).onAdShowedFullScreenContent(eq(flutterInterstitialAd));
-    verify(mockManager).onAdDismissedFullScreenContent(eq(flutterInterstitialAd));
-    verify(mockManager).onAdImpression(eq(flutterInterstitialAd));
+    verify(mockManager).onAdShowedFullScreenContent(eq(1));
+    verify(mockManager).onAdDismissedFullScreenContent(eq(1));
+    verify(mockManager).onAdImpression(eq(1));
+    assertNull(flutterInterstitialAd.getPlatformView());
   }
 
   @Test
@@ -220,6 +223,6 @@ public class FlutterInterstitialAdTest {
         .setFullScreenContentCallback(any(FullScreenContentCallback.class));
 
     flutterInterstitialAd.show();
-    verify(mockManager).onFailedToShowFullScreenContent(eq(flutterInterstitialAd), eq(adError));
+    verify(mockManager).onFailedToShowFullScreenContent(eq(1), eq(adError));
   }
 }