+++ packages/google_mobile_ads/android/src/test/java/io/flutter/plugins/googlemobileads/FlutterNativeAdTest.java
@@ -15,6 +15,8 @@
 package io.flutter.plugins.googlemobileads;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
 import static org.mockito.ArgumentCaptor.forClass;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
@@ -35,7 +37,9 @@ import com.google.android.gms.ads.admanager.AdManagerAdRequest;
 import com.google.android.gms.ads.nativead.NativeAd;
 import com.google.android.gms.ads.nativead.NativeAd.OnNativeAdLoadedListener;
 import com.google.android.gms.ads.nativead.NativeAdOptions;
+import com.google.android.gms.ads.nativead.NativeAdView;
 import io.flutter.plugin.common.BinaryMessenger;
+import io.flutter.plugin.platform.PlatformView;
 import io.flutter.plugins.googlemobileads.FlutterAd.FlutterLoadAdError;
 import io.flutter.plugins.googlemobileads.GoogleMobileAdsPlugin.NativeAdFactory;
 import java.util.Map;
@@ -67,6 +71,7 @@ public class FlutterNativeAdTest {
     Map<String, Object> mockOptions = mock(Map.class);
     final FlutterNativeAd nativeAd =
         new FlutterNativeAd(
+            1,
             testManager,
             "testId",
             mockNativeAdFactory,
@@ -137,13 +142,13 @@ public class FlutterNativeAdTest {
     verify(mockNativeAd).setOnPaidEventListener(any(FlutterPaidEventListener.class));
 
     verify(mockNativeAdFactory).createNativeAd(eq(mockNativeAd), eq(mockOptions));
-    verify(testManager).onAdOpened(eq(nativeAd));
-    verify(testManager).onAdClosed(eq(nativeAd));
-    verify(testManager).onNativeAdClicked(eq(nativeAd));
-    verify(testManager).onAdImpression(eq(nativeAd));
-    verify(testManager).onAdLoaded(eq(nativeAd), eq(responseInfo));
+    verify(testManager).onAdOpened(eq(1));
+    verify(testManager).onAdClosed(eq(1));
+    verify(testManager).onNativeAdClicked(eq(1));
+    verify(testManager).onAdImpression(eq(1));
+    verify(testManager).onAdLoaded(eq(1), eq(responseInfo));
     FlutterLoadAdError expectedError = new FlutterLoadAdError(loadAdError);
-    verify(testManager).onAdFailedToLoad(eq(nativeAd), eq(expectedError));
+    verify(testManager).onAdFailedToLoad(eq(1), eq(expectedError));
     final ArgumentCaptor<FlutterAdValue> adValueCaptor = forClass(FlutterAdValue.class);
     verify(testManager).onPaidEvent(eq(nativeAd), adValueCaptor.capture());
     assertEquals(adValueCaptor.getValue().currencyCode, "Dollars");
@@ -158,10 +163,15 @@ public class FlutterNativeAdTest {
     when(mockFlutterRequest.asAdRequest()).thenReturn(mockRequest);
     FlutterAdLoader mockLoader = mock(FlutterAdLoader.class);
     NativeAdFactory mockNativeAdFactory = mock(GoogleMobileAdsPlugin.NativeAdFactory.class);
+    NativeAdView mockNativeAdView = mock(NativeAdView.class);
+    doReturn(mockNativeAdView)
+        .when(mockNativeAdFactory)
+        .createNativeAd(any(NativeAd.class), any(Map.class));
     @SuppressWarnings("unchecked")
     Map<String, Object> mockOptions = mock(Map.class);
     final FlutterNativeAd nativeAd =
         new FlutterNativeAd(
+            1,
             testManager,
             "testId",
             mockNativeAdFactory,
@@ -215,13 +225,25 @@ public class FlutterNativeAdTest {
             eq(mockRequest));
 
     verify(mockNativeAdFactory).createNativeAd(eq(mockNativeAd), eq(mockOptions));
-    verify(testManager).onAdOpened(eq(nativeAd));
-    verify(testManager).onAdClosed(eq(nativeAd));
-    verify(testManager).onNativeAdClicked(eq(nativeAd));
-    verify(testManager).onAdImpression(eq(nativeAd));
-    verify(testManager).onAdLoaded(eq(nativeAd), eq(responseInfo));
+    verify(testManager).onAdLoaded(eq(1), eq(responseInfo));
+    verify(testManager).onAdOpened(eq(1));
+    verify(testManager).onAdClosed(eq(1));
+    verify(testManager).onNativeAdClicked(eq(1));
+    verify(testManager).onAdImpression(eq(1));
     FlutterLoadAdError expectedError = new FlutterLoadAdError(loadAdError);
-    verify(testManager).onAdFailedToLoad(eq(nativeAd), eq(expectedError));
+    verify(testManager).onAdFailedToLoad(eq(1), eq(expectedError));
+
+    // Check that platform view is defined.
+    PlatformView platformView = nativeAd.getPlatformView();
+    assertEquals(platformView.getView(), mockNativeAdView);
+    // getPlatformView() should be null after dispose() is invoked, but the platform view should
+    // still return the view.
+    nativeAd.dispose();
+    assertNull(nativeAd.getPlatformView());
+    assertNotNull(platformView.getView());
+    // Platform view's reference to the view isn't cleared until dispose() is invoked on it.
+    platformView.dispose();
+    assertNull(platformView.getView());
   }
 
   @Test(expected = IllegalStateException.class)