+++ packages/google_mobile_ads/android/src/test/java/io/flutter/plugins/googlemobileads/GoogleMobileAdsTest.java
@@ -95,6 +95,7 @@ public class GoogleMobileAdsTest {
   public void loadAd() {
     final FlutterBannerAd bannerAd =
         new FlutterBannerAd(
+            1,
             testManager,
             "testId",
             request,
@@ -110,11 +111,12 @@ public class GoogleMobileAdsTest {
   @Test
   public void disposeAd_banner() {
     FlutterBannerAd bannerAd = mock(FlutterBannerAd.class);
+
     testManager.trackAd(bannerAd, 2);
     assertNotNull(testManager.adForId(2));
     assertNotNull(testManager.adIdFor(bannerAd));
     testManager.disposeAd(2);
-    verify(bannerAd).destroy();
+    verify(bannerAd).dispose();
     assertNull(testManager.adForId(2));
     assertNull(testManager.adIdFor(bannerAd));
   }
@@ -122,11 +124,12 @@ public class GoogleMobileAdsTest {
   @Test
   public void disposeAd_adManagerBanner() {
     FlutterAdManagerBannerAd adManagerBannerAd = mock(FlutterAdManagerBannerAd.class);
+
     testManager.trackAd(adManagerBannerAd, 2);
     assertNotNull(testManager.adForId(2));
     assertNotNull(testManager.adIdFor(adManagerBannerAd));
     testManager.disposeAd(2);
-    verify(adManagerBannerAd).destroy();
+    verify(adManagerBannerAd).dispose();
     assertNull(testManager.adForId(2));
     assertNull(testManager.adIdFor(adManagerBannerAd));
   }
@@ -134,11 +137,12 @@ public class GoogleMobileAdsTest {
   @Test
   public void disposeAd_native() {
     FlutterNativeAd flutterNativeAd = mock(FlutterNativeAd.class);
+
     testManager.trackAd(flutterNativeAd, 2);
     assertNotNull(testManager.adForId(2));
     assertNotNull(testManager.adIdFor(flutterNativeAd));
     testManager.disposeAd(2);
-    verify(flutterNativeAd).destroy();
+    verify(flutterNativeAd).dispose();
     assertNull(testManager.adForId(2));
     assertNull(testManager.adIdFor(flutterNativeAd));
   }
@@ -228,6 +232,7 @@ public class GoogleMobileAdsTest {
   public void flutterAdListener_onAdLoaded() {
     final FlutterBannerAd bannerAd =
         new FlutterBannerAd(
+            0,
             testManager,
             "testId",
             request,
@@ -266,7 +271,7 @@ public class GoogleMobileAdsTest {
     doReturn("class-name").when(responseInfo).getMediationAdapterClassName();
     doReturn(adapterResponses).when(responseInfo).getAdapterResponses();
 
-    testManager.onAdLoaded(bannerAd, responseInfo);
+    testManager.onAdLoaded(0, responseInfo);
 
     final MethodCall call = getLastMethodCall();
     assertEquals("onAdEvent", call.method);
@@ -282,6 +287,7 @@ public class GoogleMobileAdsTest {
   public void flutterAdListener_onAdLoaded_responseInfoNull() {
     final FlutterBannerAd bannerAd =
         new FlutterBannerAd(
+            0,
             testManager,
             "testId",
             request,
@@ -289,7 +295,7 @@ public class GoogleMobileAdsTest {
             new BannerAdCreator(testManager.activity));
     testManager.trackAd(bannerAd, 0);
 
-    testManager.onAdLoaded(bannerAd, null);
+    testManager.onAdLoaded(0, null);
 
     final MethodCall call = getLastMethodCall();
     assertEquals("onAdEvent", call.method);
@@ -304,6 +310,7 @@ public class GoogleMobileAdsTest {
   public void flutterAdListener_onAdFailedToLoad() {
     final FlutterBannerAd bannerAd =
         new FlutterBannerAd(
+            0,
             testManager,
             "testId",
             request,
@@ -311,8 +318,7 @@ public class GoogleMobileAdsTest {
             new BannerAdCreator(testManager.activity));
     testManager.trackAd(bannerAd, 0);
 
-    testManager.onAdFailedToLoad(
-        bannerAd, new FlutterAd.FlutterLoadAdError(1, "hi", "friend", null));
+    testManager.onAdFailedToLoad(0, new FlutterAd.FlutterLoadAdError(1, "hi", "friend", null));
 
     final MethodCall call = getLastMethodCall();
     assertEquals("onAdEvent", call.method);
@@ -331,6 +337,7 @@ public class GoogleMobileAdsTest {
   public void flutterAdListener_onAppEvent() {
     final FlutterBannerAd bannerAd =
         new FlutterBannerAd(
+            0,
             testManager,
             "testId",
             request,
@@ -338,7 +345,7 @@ public class GoogleMobileAdsTest {
             new BannerAdCreator(testManager.activity));
     testManager.trackAd(bannerAd, 0);
 
-    testManager.onAppEvent(bannerAd, "color", "red");
+    testManager.onAppEvent(0, "color", "red");
 
     final MethodCall call = getLastMethodCall();
     assertEquals("onAdEvent", call.method);
@@ -356,6 +363,7 @@ public class GoogleMobileAdsTest {
   public void flutterAdListener_onAdOpened() {
     final FlutterBannerAd bannerAd =
         new FlutterBannerAd(
+            0,
             testManager,
             "testId",
             request,
@@ -363,7 +371,7 @@ public class GoogleMobileAdsTest {
             new BannerAdCreator(testManager.activity));
     testManager.trackAd(bannerAd, 0);
 
-    testManager.onAdOpened(bannerAd);
+    testManager.onAdOpened(0);
 
     final MethodCall call = getLastMethodCall();
     assertEquals("onAdEvent", call.method);
@@ -388,10 +396,11 @@ public class GoogleMobileAdsTest {
                     return null;
                   }
                 })
+            .setId(0)
             .build();
     testManager.trackAd(nativeAd, 0);
 
-    testManager.onNativeAdClicked(nativeAd);
+    testManager.onNativeAdClicked(0);
 
     final MethodCall call = getLastMethodCall();
     assertEquals("onAdEvent", call.method);
@@ -416,10 +425,11 @@ public class GoogleMobileAdsTest {
                     return null;
                   }
                 })
+            .setId(0)
             .build();
     testManager.trackAd(nativeAd, 0);
 
-    testManager.onAdImpression(nativeAd);
+    testManager.onAdImpression(0);
 
     final MethodCall call = getLastMethodCall();
     assertEquals("onAdEvent", call.method);
@@ -433,6 +443,7 @@ public class GoogleMobileAdsTest {
   public void flutterAdListener_onAdClosed() {
     final FlutterBannerAd bannerAd =
         new FlutterBannerAd(
+            0,
             testManager,
             "testId",
             request,
@@ -440,7 +451,7 @@ public class GoogleMobileAdsTest {
             new BannerAdCreator(testManager.activity));
     testManager.trackAd(bannerAd, 0);
 
-    testManager.onAdClosed(bannerAd);
+    testManager.onAdClosed(0);
 
     final MethodCall call = getLastMethodCall();
     assertEquals("onAdEvent", call.method);
@@ -454,11 +465,11 @@ public class GoogleMobileAdsTest {
   public void flutterAdListener_onRewardedAdUserEarnedReward() {
     FlutterAdLoader mockFlutterAdLoader = mock(FlutterAdLoader.class);
     final FlutterRewardedAd ad =
-        new FlutterRewardedAd(testManager, "testId", request, null, mockFlutterAdLoader);
+        new FlutterRewardedAd(0, testManager, "testId", request, null, mockFlutterAdLoader);
     testManager.trackAd(ad, 0);
 
     testManager.onRewardedAdUserEarnedReward(
-        ad, new FlutterRewardedAd.FlutterRewardItem(23, "coins"));
+        0, new FlutterRewardedAd.FlutterRewardItem(23, "coins"));
 
     final MethodCall call = getLastMethodCall();
     assertEquals("onAdEvent", call.method);
@@ -477,6 +488,7 @@ public class GoogleMobileAdsTest {
     // Set up testManager so that two ads have already been loaded and tracked.
     final FlutterRewardedAd rewarded = mock(FlutterRewardedAd.class);
     final FlutterBannerAd banner = mock(FlutterBannerAd.class);
+
     testManager.trackAd(rewarded, 0);
     testManager.trackAd(banner, 1);
 
@@ -494,7 +506,8 @@ public class GoogleMobileAdsTest {
 
     verify(testManagerSpy).disposeAllAds();
     verify(result).success(null);
-    verify(banner).destroy();
+    verify(rewarded).dispose();
+    verify(banner).dispose();
     assertNull(testManager.adForId(0));
     assertNull(testManager.adForId(1));
     assertNull(testManager.adIdFor(rewarded));